a2b: set DATCTL_{UP,DN} when new format is ready to be applied
I observed that setting DATCTL_DN=0 when a subordinate node still has
slots allocated in its registers results in an I2C write failure. This
is presumably because the DATCTL register on the main node is
auto-broadcast and the subordinate node in question rejects the value
because it doesn't agree with its other configuration.

This can be fixed by setting the value in the main node's new_structure
op, because at this point we know the slot allocation. If 0 downstream
(resp.  upstream) slots are allocated on the main node, DATCTL_DN (resp.
DATCTL_UP) can be set to 0. The same applies when setting it to 1.

I noticed this when reconfiguring my sound card, presumably due to some
race in ASoC. In general we don't have much control over the order of
codec callbacks, i.e., the main node's hw_free/shutdown might get called
before the subordinate node's hw_free/shutdown, resulting in the above
error when following the old logic.

Signed-off-by: Alvin Šipraga <alsi@bang-olufsen.dk>