a2b: add API for bus events
The A2B sound card wants to know how many A2B nodes there are so that it
can probe the correct number of codecs. Add a public API that lets
consumers subscribe to bus events, namely "discovery done". Also emit
that event when the initial discovery process is indeed done. A bus
status bitfield is also exposed and can be expanded.

Signed-off-by: Alvin Šipraga <alsi@bang-olufsen.dk>