ASoC: a2b-audio-graph-card2: add generic A2B graph card
This custom audio-garph-card2 driver supports building a sound card from
an A2B bus with an indeterminate number of nodes attached.

Note that this relies on skipping links, so a new hook has been added to
audio-graph-card2. This is squashed into this commit so that the
relevance is not lost.

Signed-off-by: Alvin Šipraga <alsi@bang-olufsen.dk>