a2b: refactor slot request API to fix bugs in RESPCYCS calculation
Emil noticed that the RESPCYCS formula was not being strictly adhered to
by the driver. It was in fact terribly wrong, and only really worked
because of our simple slot assignment. What actually needs to happen is
that the upstream and downstream slot activity has to be measured
between the nodes, i.e. between A and B transceivers.

This information is actually available from the codec, but to make
things cleaner, I have refactored the slot request API to report all
slot activity on each node as it has been configured. This has the
advantage that we can actually now perform a validation of the structure
before applying it. I even caught a mistake on MAP when I first tried
this out:

[   50.621205] a2b a2b-0: structure validation failed: downstream slot mismatch: node 0(B) sends 32 slots but node (A)1 receives 16 slots

These nice error messages should help us avoid some footguns in the
future.

Meanwhile, the slot request API has been simplified so that it doesn't
need the _pre() step. The new semantics are that the API expects that if
one slot requests slots, all of them should, and a new structure is not
applied until that happens. This is possible because now the full up-
and down-slot activity is reported in one go.

Signed-off-by: Alvin Šipraga <alsi@bang-olufsen.dk>