a2b: i2c-ad24xx: drop adapter parent hack
This was done to prevent spurious warnings from the lockdep validator,
but it also causes issues with I2C address conflict detection. While the
I2C bus of an A2B subordinate node is a discrete I2C segment that can
have address conflicts with the I2C segment that the A2B main node sits
on, this parent assignment causes the I2C core to walk up the tree and
detect conflicts anyway.

Remove and revisit this another time...

Signed-off-by: Alvin Šipraga <alsi@bang-olufsen.dk>