a2b: ad24xx-codec: fix simultaneous playback/capture
It should be possible to do playback and capture at the same time. As
such, it is necessary to detect when A2B bus data is requested from each
side, and arbitrate, such that the slot request only goes to the core
once.

This is not an ideal solution and will be improved upon later. As a
first step, new "TRX{A,B} {UP,DN}" widgets have been added to make the
routing more intuitive.

Signed-off-by: Alvin Šipraga <alsi@bang-olufsen.dk>