HACK: a2b: ad24xx-node: increase timeout waiting for PLL lock
Since the completion being waited on is completed in an IRQ handler, the
potential delay can be significantly higher than the prescribed maximum
stated in the A2B TRM. What we ought to do instead is just not rely on
the interrupt, but this can be fixed another time. For now, just make
the timeout 10 times longer.

Signed-off-by: Alvin Šipraga <alsi@bang-olufsen.dk>