a2b: move discovery into node probe
The previous logic of discovering a node before registering its
associated struct device made sense for the Shape, but it has since
presented challenges with the V302 easy access board. In a nutshell,
this logic precludes acquisition of resources such as regulators or
muxes prior to the discovery process, for to acquire these resources in
a node driver's probe function is too late, while to acquire them before
would require the logic to reside in the A2B core, which does not make
sense.

To ameliorate this problem, refactor the A2B code so that it is the node
driver that is responsible for triggering discovery through an API call.
The driver can then set up any relevant resources before making the
call, such as enabling regulators or configuring muxes.

This has some consequences in the wider design of the driver, and the
changes are not really an ideal architecture. But the constraints of the
EZ4/Shape use-case mean that for now this will have to do. Meanwhile I
have posted on the mailing list asking for help [1].

Other changes include removing the unused notifier chain, and making a
better distinction between enumeration versus discovery.

[1] https://lore.kernel.org/lkml/pr7gl7jwakze6d2no6j35h33a3an5osa6t4q3ad23q42ful3ne@24yfwin4j755/

Signed-off-by: Alvin Šipraga <alsi@bang-olufsen.dk>