a2b: granular routing for better structure application
In the previous implementation, a new structure would only get applied
when all A2B nodes on a bus had requested slots, or if all had freed
their slots. Since this request/free was made in the hw_params/hw_free
ops of the codec driver, it meant that a structure would only get
applied if all A2B codecs had their hw_params called. This is
problematic because not all of the DAIs are necessarily in the DAPM path
of a given playback or capture stream, meaning that the corresponding
node codec will not get hw_params called.

Solve this problem by moving the application of the structure into the
DAPM power management rather than at the dai_ops level. This way, any
node whose widget(s) are actually involved in the path of the stream
will trigger the application of a new structure.

The first step is to smash the rather coarse widgets that are currently
exposed by the codec into much more granular pieces. What I did was to
completely faithfully describe the internal path of A2B slots to/from
the TDM interface given the specific register settings (already exposed
as kcontrols). The result is a little baroque, since I had to introduce
some "PRE" widgets in order to make the application of autodisabled
kcontrols coherent with the overall power management.

With this in place, what happens is that the widgets are powered
correctly when opening for playback or capture. With that, the slot
request API was abandoned for a more explicit "reporting" API - nodes
now simply report the slot configuration as they have it programmed, and
the core respons in kind by triggering the application of a new
structure.

This carries a few more benefits. First of all, the structure gets
applied in pieces, and this means that capture and playback streams can
be started individually, and the structure will get applied in a reduced
state depending on the current slot requirement. Second of all the
structure can be changed via kcontrol at runtime - i.e. without closing
the card - with the caveat that user(space) should know what it is doing
so as not to screw up the order of slots. This is an optional feature
and the more static configuration of past times still works without
issue.

In my testing, I found that applying a new structure does not cause any
audio blips or drop-outs (as far as I can perceive, which is not very
well).

While at it, I took the opportunity to shorten the widget and control
names, since they were obnoxiously long and not very practical. Now the
control names reflect the actual register field names, which are more
intuitive if you are used to staring at the A2B reference manual. The
widget names are analogous.

Signed-off-by: Alvin Šipraga <alsi@bang-olufsen.dk>