a2b: ad24xx-codec: account for inverted semantics of I2S frame sync
The SND_SOC_DAIFMT_{N,I}B_{N,I}F macro refers to "normal" or "inverted"
frame sync with respect to the specific frame format (i2s, dsp_a, etc.).
While the driver logic was correct for dsp_a and dsp_b, the semantics
for i2s are different and the frame sync needs to always be inverted on
the A2B node to conform to the format. Fix.

Signed-off-by: Alvin Šipraga <alsi@bang-olufsen.dk>