a2b: ad24xx-gpio: use cache and GPIODAT to cache output values
When setting the value of a GPIO line to the same existing value,
gpiolib won't cache it and turn it into a NO-OP. Add regmap cache
support to the ad24xx-gpio driver so that such spurious writes can be
skipped. This saves a few I2C transfers in some scenarios.

Signed-off-by: Alvin Šipraga <alsi@bang-olufsen.dk>