a2b: optimize I2C peripheral transfer time on subordinate nodes
I2C peripheral access over distance requires configuring main- and
subordinate node registers appropriately before performing the
over-distance transfer on the BUS address of the main node. Since it is
quite common that consecutive transfers on the I2C adapter associated
with a subordinate node will repeatedly address the same peripheral, a
lot of redundant I2C transfers on both the MAIN and BUS addresses can be
skipped if the corresponding configuration registers are already set up
appropriately.

Optimize for this scenario by introducing proper regmap caching in the
top-level ad24xx-i2c driver and ad24xx-node driver. Use
regmap_update_bits() where appropriate, so that a
read-modify-(maybe-)write sequence is performed via the cache before
bothering to commit the same register state via an I2C transfer.
Furthermore, add caching of the last A2B_CHIP register value for each
subordinate node to keep track of whether a change in value is
necessary. In total, this allows for consecutive over-distance transfers
(as described above) to simply hit the BUS address of the main node,
minimizing the number of transfers as much as possible.

Special care has to be taken for Auto-Broadcast registers, which require
that the A2B_NODEADR.PERI bit is set to 0 before writing. This is now
handled on-demand, since writes to Auto-Broadcast registers are fairly
uncommon to begin with.

The precious register (A2B_INTTYPE) check has been hoisted out of the
node driver and into the interface driver, since the interface driver
already protects against access to this register separately for all
child regmaps. This corrects an issue which would have allowed for the
A2B_INTTYPE register getting clobbered via sysfs regmap access on the
interface driver.

Finally, correct a small inconsistency in the placement of the mutex
locking on the .i2c_xfer a2b_bus op. This is cosmetic but now aligns
things with the way it works for the .write and .read ops.

Signed-off-by: Alvin Šipraga <alsi@bang-olufsen.dk>